Move flow test helpers to separate file.
