+++ workflow/src/test/kotlin/com/r3/corda/sdk/token/workflow/TestHelpers.kt
@@ -74,77 +74,3 @@ inline fun <reified T : ContractState> assertNotHasStateAndRef(stateAndRef: Stat
         assert(!it.services.vaultService.queryBy<T>().states.contains(stateAndRef)) { "Found $stateAndRef in ${it.legalIdentity()} vault." }
     }
 }
-
-/** Create an evolvable token. */
-fun <T : EvolvableTokenType> StartedMockNode.createEvolvableToken(evolvableToken: T, notary: Party): CordaFuture<SignedTransaction> {
-    return transaction { startFlow(CreateEvolvableToken(transactionState = evolvableToken withNotary notary)) }
-}
-
-/** Update an evolvable token. */
-fun <T : EvolvableTokenType> StartedMockNode.updateEvolvableToken(old: StateAndRef<T>, new: T): CordaFuture<SignedTransaction> {
-    return transaction { startFlow(UpdateEvolvableToken(oldStateAndRef = old, newState = new)) }
-}
-
-fun <T : TokenType> StartedMockNode.issueTokens(
-        token: T,
-        issueTo: StartedMockNode,
-        notary: StartedMockNode,
-        amount: Amount<T>? = null,
-        anonymous: Boolean = true
-): CordaFuture<SignedTransaction> {
-    return transaction {
-        if (anonymous) {
-            startFlow(ConfidentialIssueFlow.Initiator(
-                    token = token,
-                    holder = issueTo.legalIdentity(),
-                    notary = notary.legalIdentity(),
-                    amount = amount
-            ))
-        } else {
-            startFlow(IssueToken.Initiator(
-                    token = token,
-                    issueTo = issueTo.legalIdentity(),
-                    notary = notary.legalIdentity(),
-                    amount = amount
-            ))
-        }
-    }
-}
-
-fun <T : TokenType> StartedMockNode.moveTokens(
-        token: T,
-        owner: StartedMockNode,
-        amount: Amount<T>? = null,
-        anonymous: Boolean = true
-): CordaFuture<SignedTransaction> {
-    return transaction {
-        if (anonymous) {
-            startFlow(ConfidentialMoveFlow.Initiator(
-                    ownedToken = token,
-                    holder = owner.legalIdentity(),
-                    amount = amount
-            ))
-        } else {
-            startFlow(MoveToken.Initiator(
-                    ownedToken = token,
-                    holder = owner.legalIdentity(),
-                    amount = amount
-            ))
-        }
-    }
-
-}
-
-fun <T : TokenType> StartedMockNode.redeemTokens(
-        token: T,
-        issuer: StartedMockNode,
-        amount: Amount<T>? = null,
-        anonymous: Boolean = true
-): CordaFuture<SignedTransaction> {
-    return startFlow(RedeemToken.InitiateRedeem(
-            ownedToken = token,
-            issuer = issuer.legalIdentity(),
-            amount = amount,
-            anonymous = anonymous
-    ))
-}
